remove default send_raw_transaction implementation
