+++ crates/optimism/rpc/src/eth/transaction.rs
@@ -53,11 +53,6 @@ where
 
         Ok(hash)
     }
-
-    #[inline]
-    fn broadcast_raw_transaction(&self, raw_tx: Bytes) {
-        self.inner.eth_api.broadcast_raw_transaction(raw_tx)
-    }
 }
 
 impl<N> LoadTransaction for OpEthApi<N>
