+++ crates/rpc/rpc/src/eth/core.rs
@@ -27,6 +27,7 @@ use reth_tasks::{
     TaskSpawner, TokioTaskExecutor,
 };
 use tokio::sync::{broadcast, Mutex};
+use tracing::trace;
 
 use crate::eth::EthTxBuilder;
 
@@ -446,8 +447,8 @@ where
     /// Broadcasts raw transaction if there are active subscribers.
     #[inline]
     pub fn broadcast_raw_transaction(&self, raw_tx: Bytes) {
-        if self.raw_tx_sender.receiver_count() > 0 {
-            let _ = self.raw_tx_sender.send(raw_tx);
+        if self.raw_tx_sender.send(raw_tx).is_err() {
+            trace!("no receivers for raw transaction");
         }
     }
 }
